+++ build_runner/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 ### Breaking Changes
 
+- There is now a whitelist of top level directories that will be used as a part
+  of the build, and other files will be ignored. For now those directories
+  include 'benchmark', 'bin', 'example', 'lib', 'test', 'tool', and 'web'.
+  - If this breaks your workflow please file an issue and we can look at either
+    adding additional directories or making the list configurable per project.
 - Remove `PackageGraph.orderedPackages` and `PackageGraph.dependentsOf`.
 
 The following changes are technically breaking but should not impact most